add diary overview when clicking on a day update meal overview, so the delete button is only available on the actual date of the diary entry refactoring

+++ frontend/src/components/cards/FoodItemCard.tsx
@@ -2,15 +2,18 @@ import {FoodItem} from "../../types/Diary.ts";
 import DeleteButton from "../svg/DeleteButton.tsx";
 
 type FoodItemCardProps = {
-    deleteFoodItem : (foodItemToDelete : FoodItem) => void,
-    foodItem : FoodItem
+    deleteFoodItem? : (foodItemToDelete : FoodItem) => void,
+    foodItem : FoodItem,
+    isHomescreen: boolean
 }
 export default  function FoodItemCard(props: Readonly<FoodItemCardProps>) {
 
     function onClick(){
-        props.deleteFoodItem(props.foodItem)
-
+        if(props.deleteFoodItem){
+            props.deleteFoodItem(props.foodItem)
+        }
     }
+
     return (
         <div className={"card"}>
             <div className={"card-header"}>
@@ -23,7 +26,7 @@ export default  function FoodItemCard(props: Readonly<FoodItemCardProps>) {
                             id={"servingsize"}>{" (" + props.foodItem.amount + " g)"}</span>
                     </div>
                 </div>
-                <button onClick={onClick}><DeleteButton/></button>
+                {props.deleteFoodItem && <button onClick={onClick}><DeleteButton/></button>}
             </div>
         </div>
     );
